pyproject.toml
[build-system]
requires = ["setuptools"]

build-backend = "setuptools.build_meta"

[project]
name = "Ihorium Trade-bot"

version = "1.4.1"

readme = "README.md"

requires-python = ">=3.10"

license = { file = "LICENSE.txt" }

authors = [{ name = "Zemphix", email = "zemphix@gmail.com" }]

maintainers = [{ name = "Uzgramor", email = "None" }]

dependencies = [
    "asarPy==1.0.1",
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "idna==3.8",
    "numpy==2.1.1",
    "pandas==2.2.3",
    "pybit==5.8.0",
    "pycryptodome==3.20.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-magic==0.4.27",
    "pytz==2024.2",
    "requests==2.32.3",
    "six==1.16.0",
    "ta==0.11.0",
    "tzdata==2024.1",
    "urllib3==2.2.2",
    "websocket-client==1.8.0",
    "openpyxl==3.1.5"
]
